requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
pillow>=10.0.0
